+++ app/src/main/java/com/victor_sml/playlistmaker/TracksResponse.kt
@@ -1,3 +1,5 @@
 package com.victor_sml.playlistmaker
 
+import com.victor_sml.playlistmaker.domain.Track
+
 class TracksResponse(val results: ArrayList<Track>)
