+++ app/src/main/java/com/victor_sml/playlistmaker/domain/Track.kt
@@ -1,4 +1,4 @@
-package com.victor_sml.playlistmaker
+package com.victor_sml.playlistmaker.domain
 
 import android.os.Parcelable
 import com.google.gson.TypeAdapter
