+++ app/src/main/java/com/victor_sml/playlistmaker/presentation/ui/settings/SettingsActivity.kt
@@ -1,4 +1,4 @@
-package com.victor_sml.playlistmaker
+package com.victor_sml.playlistmaker.presentation.ui.settings
 
 import android.content.Intent
 import android.net.Uri
@@ -7,6 +7,9 @@ import android.widget.FrameLayout
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.widget.Toolbar
 import com.google.android.material.switchmaterial.SwitchMaterial
+import com.victor_sml.playlistmaker.App
+import com.victor_sml.playlistmaker.DARK_THEME_ENABLED
+import com.victor_sml.playlistmaker.R
 
 class SettingsActivity : AppCompatActivity() {
     private lateinit var app: App
