+++ app/src/main/java/com/victor_sml/playlistmaker/recycler/TrackAdapter.kt
@@ -2,7 +2,7 @@ package com.victor_sml.playlistmaker.recycler
 
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import com.victor_sml.playlistmaker.Track
+import com.victor_sml.playlistmaker.domain.Track
 
 class TrackAdapter(private val delegates: List<AdapterDelegate>) :
     RecyclerView.Adapter<RecyclerView.ViewHolder>() {
