+++ app/src/main/java/com/victor_sml/playlistmaker/recycler/TrackDelegate.kt
@@ -8,9 +8,9 @@ import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import com.bumptech.glide.load.resource.bitmap.RoundedCorners
 import com.victor_sml.playlistmaker.R
-import com.victor_sml.playlistmaker.Track
+import com.victor_sml.playlistmaker.domain.Track
 import com.victor_sml.playlistmaker.TracksHandler
-import com.victor_sml.playlistmaker.utils.dpToPx
+import com.victor_sml.playlistmaker.Utils.dpToPx
 
 class TrackDelegate(
     private val tracksHandler: TracksHandler,
