--- app/src/main/java/com/victor_sml/playlistmaker/utils/DipConverter.kt
@@ -1,16 +1,0 @@
-package com.victor_sml.playlistmaker.utils
-
-import android.content.Context
-import android.util.TypedValue
-
-/**
- * Конвертирует dp в px
- */
-fun dpToPx(dp: Int, context: Context): Int {
-    val resources = context.resources
-    return TypedValue.applyDimension(
-        TypedValue.COMPLEX_UNIT_DIP,
-        dp.toFloat(),
-        resources.displayMetrics
-    ).toInt()
-}
